+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Next]
+- Fix reversed result in refined `#<=>`.
 - Remove `#to_str`, as VectorNumber is not a String-like object.
 - Due to the above, `Kernel.BigDecimal` no longer works without refinements.